refactor(symbols): вынести построение Annotation из AST в util Annotations
Разбор аннотаций жил приватно в MethodSymbolComputer. Выношу в
переиспользуемый Annotations, чтобы тот же код читал аннотации над
переменными модуля (нужно для вывода типов внедряемых зависимостей ОСени).
